ui dialog error fix: order overflow

+++ ceammc/ext/src/lib/cicm/Sources/edefine.h
@@ -615,7 +615,7 @@ struct t_eattr {
     t_float minimum; /*!< The minimum value of the attribute. */
     t_float maximum; /*!< The maximum value of the attribute. */
     t_float step; /*!< The increment or decrement step calue of the attribute. */
-    std::int8_t order; /*!< The dummy order of the attribute. */
+    std::int16_t order; /*!< The dummy order of the attribute. */
     eclip_flags clipped; /*!< If the attribute is clipped if it's value or an array of numerical values. */
     bool save; /*!< If the attribute should be saved. */
     bool paint; /*!< If the attribute should repaint the t_ebox when it has changed. */
